refactor: the layer rows are a component, not a render prop
App built the rows of every layer tree and handed the function to CutLayerPanel
as a prop, with a comment saying it had to stay in App because the drag state
lives there. Drag state is a prop like any other. What the render prop actually
bought was that nobody had to write the row's inputs down.

There are twenty of them. They are written down now, in `ui/LayerRows.jsx`, and
they travel as one `layerRows` object rather than twenty names threaded one by
one through a panel that uses none of them - the same grouping usePlayback uses,
for the same reason. LayerThumbnail goes with them, since nothing else drew one.

Fifteen imports in App were left with nothing to import for; they go too. One of
those was a second, standalone import of a name the big block already had.

App.jsx: 3781 -> 3695 lines.
